Make the move properly
