+++ webapp/src/features/event/lineup/EventLineup.tsx
@@ -7,7 +7,7 @@ import { DocumentReference } from "firebase/firestore";
 import { EventDjSignups } from "./EventSignupList";
 import { v4 as uuidv4 } from 'uuid';
 import { useEventDjCache } from "../../../contexts/useEventDjCache";
-import { AddOrCreateDjModal } from "./Components/AddOrCreateDjModal";
+import { AddOrCreateDjModal } from "./components/AddOrCreateDjModal";
 import { updateSignupForEvent } from "../util";
 
 const EventLineup = () => {
